requirements.txt
Flask==3.1.1
numpy==2.2.6
openai==1.84.0
python-dotenv==1.1.0
upstash_vector==0.8.0
google-genai==1.19.0
gunicorn==23.0.0